\name{influence.SEM-package}
\alias{influence.SEM-package}
\alias{influence.SEM}
\docType{package}
\title{
Case Influence in Structural Equation Models
}
\description{
influence.SEM provides a set of tools for evaluating several measures of
case influence for Structural Equation Models following the approach of Pek
and MacCallum (2011). It analyses models that were estimated using the function \code{sem} of the package \code{lavaan}. 
}
\details{
\tabular{ll}{
Package: \tab influence.SEM\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Depends: \tab lavaan\cr
Suggests: \tab tcltk\cr
Date: \tab 2013-10-03\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Massimiliano Pastore & Gianmarco Altoe'

Maintainer: Massimiliano Pastore <massimiliano.pastore@unipd.it>
}
\references{
Pek, J., MacCallum, R.C. (2011). Sensitivity Analysis in Structural Equation Models: Cases and Their Influence. \emph{Multivariate Behavioral Research}, 46, 202-228. 

Rosseel, Y. (2013). The \code{lavaan} tutorial. URL: \url{http://lavaan.ugent.be/tutorial/tutorial.pdf}.

Tukey, J. W. (1977). \emph{Exploratory data analysis}. Reading, MA: Addison-Wesley. 
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%###
%}
